requirements.txt
flask
torch
torchaudio
transformers
coqui-ai-tts
pydub